% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_data}
\alias{download_data}
\title{Download AutoSeg Data}
\usage{
download_data(periods, dest_dir, keep_zip = FALSE)
}
\arguments{
\item{periods}{a character vector of time periods to download}

\item{dest_dir}{a folder to save downloaded data to. Will be created if
folder does not exist}

\item{keep_zip}{if FALSE (default) the downloaded zip files are deleted after
they are uncompressed}
}
\description{
Download AutoSeg Data
}
