% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_file}
\alias{download_file}
\title{Download Aut0Seg File from URL}
\usage{
download_file(x, dest_dir, keep_zip = FALSE)
}
\arguments{
\item{x}{url to download the file from}

\item{dest_dir}{folder to save downloaded file to}

\item{keep_zip}{if FALSE (default) the downloaded zip file will be deleted
after it has been uncompressed}
}
\description{
Download Aut0Seg File from URL
}
