% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_autoseg}
\alias{read_autoseg}
\title{Read Autoseg File}
\usage{
read_autoseg(file, ...)
}
\arguments{
\item{file}{Path to file.}

\item{...}{Additional arguments passed to `readr::read_delim()`.}
}
\description{
Read Autoseg File
}
